Move data to parent package, as it's not confined to repositories
