+++ app/src/androidTest/java/ch/epfl/sweng/eventmanager/mock/repository/MockConcertRepository.java
@@ -2,7 +2,7 @@ package ch.epfl.sweng.eventmanager.mock.repository;
 
 import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.MutableLiveData;
-import ch.epfl.sweng.eventmanager.repository.data.ScheduledItem;
+import ch.epfl.sweng.eventmanager.data.ScheduledItem;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
