+++ app/src/androidTest/java/ch/epfl/sweng/eventmanager/repository/JoinedEventTestUtils.java
@@ -1,6 +1,6 @@
 package ch.epfl.sweng.eventmanager.repository;
 
-import ch.epfl.sweng.eventmanager.repository.data.JoinedEvent;
+import ch.epfl.sweng.eventmanager.data.JoinedEvent;
 import ch.epfl.sweng.eventmanager.repository.room.daos.JoinedEventDao;
 import java.util.ArrayList;
 import java.util.Collections;
