+++ app/src/androidTest/java/ch/epfl/sweng/eventmanager/repository/JoinedScheduleItemRepositoryTest.java
@@ -2,7 +2,7 @@ package ch.epfl.sweng.eventmanager.repository;
 
 import android.arch.core.executor.testing.InstantTaskExecutorRule;
 import android.support.test.runner.AndroidJUnit4;
-import ch.epfl.sweng.eventmanager.repository.data.JoinedScheduleItem;
+import ch.epfl.sweng.eventmanager.data.JoinedScheduleItem;
 import ch.epfl.sweng.eventmanager.repository.room.Converters;
 import org.junit.Before;
 import org.junit.Rule;
