+++ app/src/androidTest/java/ch/epfl/sweng/eventmanager/repository/room/JoinedScheduleItemDao.java
@@ -4,8 +4,7 @@ import android.arch.core.executor.testing.InstantTaskExecutorRule;
 import android.support.test.runner.AndroidJUnit4;
 import ch.epfl.sweng.eventmanager.repository.JoinedScheduleItemTestUtils;
 import ch.epfl.sweng.eventmanager.repository.LiveDataTestUtil;
-import ch.epfl.sweng.eventmanager.repository.data.JoinedEvent;
-import ch.epfl.sweng.eventmanager.repository.data.JoinedScheduleItem;
+import ch.epfl.sweng.eventmanager.data.JoinedScheduleItem;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestRule;
