+++ app/src/main/java/ch/epfl/sweng/eventmanager/data/JoinedEvent.java
@@ -1,13 +1,9 @@
-package ch.epfl.sweng.eventmanager.repository.data;
+package ch.epfl.sweng.eventmanager.data;
 
 import android.arch.persistence.room.ColumnInfo;
 import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.PrimaryKey;
-import android.arch.persistence.room.TypeConverter;
 import android.support.annotation.NonNull;
-
-import java.util.Objects;
-import java.util.UUID;
 
 /**
  * Defines the JoinedEvent Entity. We only store the event-id and the name of the event to minimise storage space
