+++ app/src/main/java/ch/epfl/sweng/eventmanager/data/package-info.java
@@ -3,4 +3,4 @@
  *
  * @author Louis Vialar
  */
-package ch.epfl.sweng.eventmanager.repository.data;
+package ch.epfl.sweng.eventmanager.data;
