+++ app/src/main/java/ch/epfl/sweng/eventmanager/repository/room/daos/JoinedEventDao.java
@@ -2,7 +2,7 @@ package ch.epfl.sweng.eventmanager.repository.room.daos;
 
 import android.arch.lifecycle.LiveData;
 import android.arch.persistence.room.*;
-import ch.epfl.sweng.eventmanager.repository.data.JoinedEvent;
+import ch.epfl.sweng.eventmanager.data.JoinedEvent;
 
 import java.util.List;
