+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/eventShowcase/fragments/schedule/MyScheduleFragment.java
@@ -13,7 +13,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import ch.epfl.sweng.eventmanager.R;
-import ch.epfl.sweng.eventmanager.repository.data.ScheduledItem;
+import ch.epfl.sweng.eventmanager.data.ScheduledItem;
 
 import java.io.File;
 import java.io.FileOutputStream;
