+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/userManager/LoginActivity.java
@@ -3,9 +3,7 @@ package ch.epfl.sweng.eventmanager.ui.userManager;
 import android.content.Context;
 import android.content.Intent;
 import android.os.AsyncTask;
-import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.RequiresApi;
 import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.view.View;
@@ -15,12 +13,11 @@ import android.widget.EditText;
 import android.widget.ProgressBar;
 import ch.epfl.sweng.eventmanager.R;
 import ch.epfl.sweng.eventmanager.repository.UserRepository;
-import ch.epfl.sweng.eventmanager.repository.data.User;
+import ch.epfl.sweng.eventmanager.data.User;
 import ch.epfl.sweng.eventmanager.userManagement.Session;
 import dagger.android.AndroidInjection;
 
 import javax.inject.Inject;
-import java.util.Optional;
 
 /**
  * A login screen that offers login via email/password.
