+++ app/src/test/java/ch/epfl/sweng/eventmanager/data/PositionTest.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.eventmanager.repository.data;
+package ch.epfl.sweng.eventmanager.data;
 
 import com.google.android.gms.maps.model.LatLng;
 import org.junit.Test;
