+++ app/src/test/java/ch/epfl/sweng/eventmanager/data/UserTest.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.eventmanager.repository.data;
+package ch.epfl.sweng.eventmanager.data;
 
 import org.junit.Test;
